+++ app/components/dashboard/Category.tsx
@@ -76,7 +76,7 @@ interface Props {
 
 export const CategoryItem = ({category, borderColor}: Props) => {
     const router = useRouter();
-    const {t} = useTranslation();
+    const { t } = useTranslation();
 
     const conatinerStyle = {
         justifyContent: "center",
@@ -94,7 +94,7 @@ export const CategoryItem = ({category, borderColor}: Props) => {
     }
 
     return (<Grid container xs={12} sm={2} sx={conatinerStyle}>
-        <CustomH3>{t(category.label)}</CustomH3>
+        <CustomH3>{t(`abc-${category.label.toLocaleLowerCase().replace(" ", "-")}`)}</CustomH3>
         <Paragraph>
             {t(category.details)}
         </Paragraph>
