+++ app/helpers/text.ts
@@ -1,0 +1,10 @@
+export const prepareValueForTranslation = (value: string) => {
+    let newValue;
+
+    newValue = value.toLowerCase()
+            .replace(/\s+/g, "-")
+            .replace(/\s/g, "")
+            .replace(/-\([^)]*\)$/, "");
+
+    return `abc-${newValue}`;
+}
