+++ app/locales/fr/common.json
@@ -122,5 +122,6 @@
   "abc-school-address": "Adresse d'établissement",
   "abc-update-school": "Modifier les informations de l'établissement",
   "abc-school-update-successfully": "Les informations sur l'établissement ont été mises à jour avec succès",
-  "abc-school-update-error": "Échec de la mise à jour des informations sur l'établissement"
+  "abc-school-update-error": "Échec de la mise à jour des informations sur l'établissement",
+  "abc-confirm-duplicate": "Une donnée d'activité existe déjà pour {{type}}, voulez-vous quand même enregistrer une nouvelle donnée ?"
 }
